fix some datatypes
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@23500 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/httpd/visuals/site_survey.c
@@ -206,7 +206,7 @@ void ej_dump_site_survey(webs_t wp, int argc, char_t ** argv)
 			: live_translate("share.yes"), 31);
 
 		char *netmode;
-		long netmodecap = site_survey_lists[i].capability;
+		int netmodecap = site_survey_lists[i].capability;
 
 		netmodecap &= (DOT11_CAP_ESS | DOT11_CAP_IBSS);
 		if (netmodecap == DOT11_CAP_ESS)
