+++ src/router/shared/linux_osl.h
@@ -30,6 +30,6 @@
 #include <time.h>
 #include <stdarg.h>
 
-int osl_join_multicast(struct iface *pif, int fd, ulong ipaddr, ushort port);	// added 
+int osl_join_multicast(struct iface *pif, int fd, unsigned int ipaddr, ushort port);	// added 
 										// - 
 										// tofu
